refactor: Rename arrays.ts to utils.ts
